chore(release): releasing 1.0.31
What's new in 1.0.31?
----
- Added model service to get the data model

The model exposes all of the following:

- Event types
- Timeseries
- Object types
- Object attributes
- Owner attributes
- Sessionizers

The returned model contains everything that is applied the zone (sandbox or production) your are currently working with. By definition, everything that is available in the production view is also available in the sandbox view. The opposite is not true.

[skip ci]